requirements.txt
openai>=1.0.0
anthropic>=0.3.0
tiktoken>=0.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.3
langchain>=0.1.0
langchain-openai>=0.0.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tenacity>=8.2.0
PyPDF2>=3.0.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
llama-parse>=0.4.0
llama-cloud-services>=0.1.0
python-dotenv>=1.0.0
